requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-ollama
langchain-groq
langchain-chroma
langchain-huggingface
sentence-transformers
langchain-aws
fastapi
streamlit
python-dotenv
faiss-cpu
chromadb
pypdf
arxiv
wikipedia
duckduckgo-search
validators==0.28.1
youtube_transcript_api
unstructured
pytube
datasets
tiktoken
gradio
boto3
awscli
crewai
crewai_tools
mangum
uvicorn